Move board to own file, import to main file
